refactor: move yaml internal transformations to printer.preprocess
